+++ versions/1.20.4-neoforge/src/main/java/dev/yurisuika/compost/world/level/storage/loot/predicates/MatchCompostable.java
@@ -42,7 +42,6 @@ public record MatchCompostable(Optional<ItemPredicate> predicate) implements Loo
             }
         }
         return false;
-
     }
 
     public static Builder compostableMatches(ItemPredicate.Builder builder) {