+++ versions/1.21.5-neoforge/src/main/java/dev/yurisuika/compost/mixin/minecraft/world/level/block/BlocksMixin.java
@@ -1,5 +1,6 @@
 package dev.yurisuika.compost.mixin.minecraft.world.level.block;
 
+import dev.yurisuika.compost.Compost;
 import dev.yurisuika.compost.world.level.block.ContainerComposterBlock;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
@@ -16,6 +17,7 @@ public abstract class BlocksMixin {
 
     @ModifyArg(method = "<clinit>", slice = @Slice(from = @At(value = "CONSTANT", args = "stringValue=composter")), at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/Blocks;register(Ljava/lang/String;Ljava/util/function/Function;Lnet/minecraft/world/level/block/state/BlockBehaviour$Properties;)Lnet/minecraft/world/level/block/Block;", ordinal = 0), index = 1)
     private static Function<Properties, Block> replaceComposter(Function<Properties, Block> function) {
+        Compost.LOGGER.info("Replacing the vanilla composter with a Compost composter!");
         return ContainerComposterBlock::new;
     }