+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 - Fixed NeoForge event bus not being defined crashing 1.21.2 through 1.21.5.
 - Fixed mod failing to load due to datagen entrypoint for NeoForge 21.4 not being compatible with 21.2/3.
+- Added warning in log when a composter attempts to fill from a block entity with an empty compostables list (such as using a give command with the block state of "level" set to "7").